+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Feature) Optional Assertions
 - (Feature) Deprecate Actions
 - (Bugfix) Debug mode
+- (Improvement) Switch to Lease API
 
 ## [1.2.32](https://github.com/arangodb/kube-arangodb/tree/1.2.32) (2023-08-07)
 - (Feature) Backup lifetime - remove Backup once its lifetime has been reached